+++ js/data.js
@@ -1,0 +1,34 @@
+const OBJECT_VALUE = 25;
+const PHOTO_VALUE = 25;
+const LIKES_VALUE__MIN = 15;
+const LIKES_VALUE__MAX = 200;
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Никита',
+  'Рита',
+  'Олег',
+  'Анна',
+  'Изабелла',
+  'Илья',
+  'Сема'
+];
+
+const DESC_PHOTO = [
+  'шикарное фото',
+  'замечательное фото',
+  'свежее фото',
+  'удивительное фото',
+  'классное фото',
+  'вау фото'
+];
+
+export { OBJECT_VALUE, PHOTO_VALUE, LIKES_VALUE__MIN, LIKES_VALUE__MAX, MESSAGES, NAMES, DESC_PHOTO };
