+++ js/main.js
@@ -1,80 +1,5 @@
-const OBJECT_VALUE = 25;
-const PHOTO_VALUE = 25;
-const LIKES_VALUE__MIN = 15;
-const LIKES_VALUE__MAX = 200;
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const NAMES = [
-  'Никита',
-  'Рита',
-  'Олег',
-  'Анна',
-  'Изабелла',
-  'Илья',
-  'Сема'
-];
-
-const DESC_PHOTO = [
-  'шикарное фото',
-  'замечательное фото',
-  'свежее фото',
-  'удивительное фото',
-  'классное фото',
-  'вау фото'
-];
-
-const getRandomNumber = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (elements) => elements[getRandomNumber(0, elements.length - 1)];
-
-const createRandomValue = (min, max) => {
-  const previousValues = [];
-
-  return function () {
-    let currentValue = getRandomNumber(min, max);
-    if (previousValues.length >= (max - min + 1)) {
-      return null;
-    }
-    while (previousValues.includes(currentValue)) {
-      currentValue = getRandomNumber(min, max);
-    }
-    previousValues.push(currentValue);
-    return currentValue;
-  };
-};
-
-const getPhotoId = createRandomValue(1, OBJECT_VALUE);
-const getPhotoSrc = createRandomValue(1, PHOTO_VALUE);
-const getCommentId = createRandomValue(1, 500);
-
-const createComments = () => ({
-  id: getCommentId(),
-  avatar: `img/avatar-${createRandomValue(1, 6)()}.svg`,
-  message: getRandomArrayElement(MESSAGES),
-  name: getRandomArrayElement(NAMES),
-});
-
-
-const photoCard = () => ({
-  id: getPhotoId(),
-  url: `photos/${getPhotoSrc()}.jpg`,
-  description: getRandomArrayElement(DESC_PHOTO),
-  likes: getRandomNumber(LIKES_VALUE__MIN, LIKES_VALUE__MAX),
-  comments: Array.from({ length: getRandomNumber(0, 30) }, createComments),
-});
+import { photoCard } from './create-element';
 
 const allCards = Array.from({ length: 25 }, photoCard);
 
+console.log(allCards);
